requirements.txt
google-genai==1.25.0
python-dotenv==1.1.1
pydantic==2.11.7
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4
